+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.2.0 - 2026-08-09
+
+### Changed
+
+- Improve commit behavior
+- Move gg commit conventions from gg_git to gg_one_core
+- Move the git and process plumbing to gg_git
+- Record the doCommit state in system commits again
+
 ## 2.1.0 - 2026-08-09
 
 ## 2.0.0 - 2026-08-08